packages/backend/requirements.txt
# -----------------------------
# Core framework
# -----------------------------
Django==5.0.6
asgiref==3.8.1
sqlparse==0.4.2
tzdata==2022.7

# -----------------------------
# REST / Auth
# -----------------------------
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
drf-access-policy==1.3.0
PyJWT==2.4.0

# -----------------------------
# Channels / ASGI
# -----------------------------
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
uvicorn==0.27.1
uvloop==0.19.0
websockets==12.0

# -----------------------------
# GraphQL
# -----------------------------
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
django-channels-graphql-ws @ https://github.com/apptension/DjangoChannelsGraphqlWs/archive/refs/tags/v1.0.0rc7.tar.gz#sha256=90cbb41958458624eb18accbbebd6489c51f43e63785430d0ea1938b7826f9c6
graphene-file-upload==1.3.0

# -----------------------------
# Celery / Async
# -----------------------------
celery==5.4.0
kombu==5.3.7
billiard==4.2.0
amqp==5.2.0
vine==5.1.0
flower==2.0.1

# -----------------------------
# Redis / Cache
# -----------------------------
redis==5.0.1
django-redis==5.4.0

# -----------------------------
# AWS
# -----------------------------
boto3==1.26.58
aws-xray-sdk==2.11.0
moto==4.1.1
cfn-lint==0.61.3
aws-sam-translator==1.47.0
django-ses==4.1.0

# -----------------------------
# Storage / Static
# -----------------------------
django-storages==1.13.2
whitenoise==6.3.0
Pillow==9.4.0

# -----------------------------
# Auth / Social
# -----------------------------
social-auth-app-django==5.0.0
social-auth-core==4.3.0
python3-openid==3.2.0
oauthlib==3.2.0
requests-oauthlib==1.3.1

# -----------------------------
# Security / Crypto
# -----------------------------
cryptography==42.0.0
pyOpenSSL==24.0.0
pyotp==2.8.0
python-jose==3.3.0
ecdsa==0.18.0

# -----------------------------
# HTTP
# -----------------------------
requests==2.28.2
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.8.4
anyio==4.2.0

# -----------------------------
# Utils
# -----------------------------
python-dotenv==0.20.0
python-dateutil==2.8.2
humanize==4.9.0
inflection==0.5.1
hashids==1.3.1
jsonschema==3.2.0
MarkupSafe==2.1.1
Jinja2==3.1.2

# -----------------------------
# DB
# -----------------------------
psycopg2-binary==2.9.9

# -----------------------------
# Monitoring
# -----------------------------
sentry-sdk

# -----------------------------
# Dev / Test
# -----------------------------
black==23.1.0
ruff==0.0.261
pytest==8.2.1
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-dotenv==0.5.2
pytest-mock==3.10.0
coverage==7.1.0
factory-boy==3.2.1
Faker==13.15.0
freezegun==1.2.2

# -----------------------------
# Docs / Tooling
# -----------------------------
rich==14.2.0
watchfiles==0.21.0
watchdog==2.3.1

dj-stripe==2.8.1
django-hosts==5.2
coreapi==2.3.3
django-hashid-field==3.3.7
django-ratelimit==4.0.0
openai==0.27.2
openpyxl==3.1.5
gunicorn
django-environ
